Move TimeServiceBackend back to the API server.
