+++ ledger/participant-state/kvutils/app/src/main/scala/com/daml/ledger/participant/state/kvutils/app/LedgerFactory.scala
@@ -5,10 +5,9 @@ package com.daml.ledger.participant.state.kvutils.app
 
 import akka.stream.Materializer
 import com.codahale.metrics.{MetricRegistry, SharedMetricRegistries}
-import com.daml.ledger.participant.state.v1.TimeServiceBackend
 import com.digitalasset.ledger.api.auth.{AuthService, AuthServiceWildcard}
 import com.digitalasset.logging.LoggingContext
-import com.digitalasset.platform.apiserver.ApiServerConfig
+import com.digitalasset.platform.apiserver.{ApiServerConfig, TimeServiceBackend}
 import com.digitalasset.platform.indexer.{IndexerConfig, IndexerStartupMode}
 import com.digitalasset.resources.ResourceOwner
 import scopt.OptionParser
