+++ ledger/sandbox/src/main/scala/com/digitalasset/platform/apiserver/TimeServiceBackend.scala
@@ -1,7 +1,7 @@
 // Copyright (c) 2020 The DAML Authors. All rights reserved.
 // SPDX-License-Identifier: Apache-2.0
 
-package com.daml.ledger.participant.state.v1
+package com.digitalasset.platform.apiserver
 
 import java.time.Instant
 import java.util.concurrent.atomic.AtomicReference
