+++ ledger/sandbox/src/main/scala/com/digitalasset/platform/apiserver/services/ApiTimeService.scala
@@ -8,7 +8,6 @@ import java.time.Instant
 import akka.NotUsed
 import akka.stream.Materializer
 import akka.stream.scaladsl.Source
-import com.daml.ledger.participant.state.v1.TimeServiceBackend
 import com.digitalasset.api.util.TimestampConversion._
 import com.digitalasset.dec.DirectExecutionContext
 import com.digitalasset.grpc.adapter.ExecutionSequencerFactory
@@ -18,6 +17,7 @@ import com.digitalasset.ledger.api.v1.testing.time_service._
 import com.digitalasset.logging.{ContextualizedLogger, LoggingContext}
 import com.digitalasset.platform.akkastreams.dispatcher.SignalDispatcher
 import com.digitalasset.platform.api.grpc.GrpcApiService
+import com.digitalasset.platform.apiserver.TimeServiceBackend
 import com.digitalasset.platform.server.api.validation.FieldValidations
 import com.google.protobuf.empty.Empty
 import io.grpc.{ServerServiceDefinition, Status, StatusRuntimeException}
